let library module can rename by setting

+++ src/language/common/commonLanguage.ts
@@ -6,22 +6,29 @@ import LineReader from 'n-readlines-next'
 import * as fsWalk from '@nodelib/fs.walk'
 import * as path from 'path'
 
-function replaceTextLineByLineAtPath(replacePath: string, from: string, to: string) {
-  const liner = new LineReader(replacePath)
+export const replaceTextLineByLineAtPath = (
+  replacePath: string, from: string, to: string,
+  encoding?: 'utf-8'
+): void => {
+  const liner = new LineReader(replacePath, { readChunk: 1024, newLineCharacter: '\n' })
   const lineCache = []
   let line = liner.next()
+  let mustReplace = false
   do {
-    const lS = line.toString('utf-8')
+    const lS = line.toString(encoding)
     if (lS.search(from) !== -1) {
       lineCache.push(lS.replace(from, to))
+      mustReplace = true
     } else {
       lineCache.push(lS)
     }
     line = liner.next()
   } while (line)
-  const newFileContent = lineCache.join('\n')
-  // logDebug(`newFileContent:\n${newFileContent}`)
-  fsExtra.writeFileSync(replacePath, newFileContent)
+  if (mustReplace) {
+    const newFileContent = lineCache.join('\n')
+    // logDebug(`newFileContent:\n${newFileContent}`)
+    fsExtra.writeFileSync(replacePath, newFileContent)
+  }
 }
 
 export const replaceTextByPathList = (from: string, to: string, ...pathList: string[]): void => {
